+++ pmfp/entrypoint/grpc/build_/core.py
@@ -24,15 +24,15 @@ class Build(EndPoint):
                 "type": "string",
                 "enum": ["py", "cython", "go", "js", "CXX"]
             },
-            "as_type": {
-                "type": "array",
-                "title": "a",
-                "items": {
-                    "type": "string",
-                    "enum": ["serv", "cli", "aio", "nogen", "mp"]
-                },
-                "description": "目的,为空时不会生成代码,当有`serv`时会生成服务端代码;当有`cli`时会生成客户端代码;当有`nogen`只是将proto文件复制到目标文件夹"
-            },
+            # "as_type": {
+            #     "type": "array",
+            #     "title": "a",
+            #     "items": {
+            #         "type": "string",
+            #         "enum": ["serv", "cli", "aio", "nogen", "mp"]
+            #     },
+            #     "description": "目的,为空时不会生成代码,当有`serv`时会生成服务端代码;当有`cli`时会生成客户端代码;当有`nogen`只是将proto文件复制到目标文件夹"
+            # },
             "to": {
                 "type": "string",
                 "title": "t",